+++ app/src/main/jni/vpnproxy-jni/vpnproxy.c
@@ -24,7 +24,7 @@
 #include "utils.c"
 #include "vpnproxy.h"
 #include "pcap.h"
-#include "../../../../../../nDPI/src/include/ndpi_protocol_ids.h"
+#include "ndpi_protocol_ids.h"
 
 #define CAPTURE_STATS_UPDATE_FREQUENCY_MS 300
 #define CONNECTION_DUMP_UPDATE_FREQUENCY_MS 3000
@@ -79,6 +79,7 @@ typedef struct jni_classes {
 static jni_classes_t cls;
 static jni_methods_t mids;
 static bool running = false;
+static bool dump_connections_now = false;
 
 /* TCP/IP packet to hold the mitmproxy header */
 static char mitmproxy_pkt_buffer[] = {
@@ -299,11 +300,18 @@ static void javaPcapDump(zdtun_t *tun, vpnproxy_data_t *proxy) {
 
 /* ******************************************************* */
 
-static bool shouldIgnoreApp(vpnproxy_data_t *proxy, int uid) {
+static bool shouldIgnoreConn(vpnproxy_data_t *proxy, const zdtun_5tuple_t *tuple, const conn_data_t *data) {
+#if 0
+    int uid = data.uid;
     bool is_unknown_app = ((uid == -1) || (uid == 1051 /* netd DNS resolver */));
 
     if(((proxy->uid_filter != -1) && (proxy->uid_filter != uid))
         && (!is_unknown_app || !proxy->capture_unknown_app_traffic))
+        return true;
+#endif
+
+    // ignore some internal communications, e.g. DNS-over-TLS check on port 853
+    if((tuple->dst_ip == proxy->vpn_dns) && (ntohs(tuple->dst_port) != 53))
         return true;
 
     return false;
@@ -344,7 +352,7 @@ static void account_packet(zdtun_t *tun, const char *packet, int size, uint8_t f
     if(data->ndpi_flow)
         process_ndpi_packet(data, proxy, packet, size, from_tap);
 
-    if(shouldIgnoreApp(proxy, data->uid)) {
+    if(shouldIgnoreConn(proxy, zdtun_conn_get_5tuple(conn_info), data)) {
         //log_android(ANDROID_LOG_DEBUG, "Ignoring connection: UID=%d [filter=%d]", data->uid, proxy->uid_filter);
         return;
     }
@@ -570,7 +578,7 @@ static int check_dns_req_dnat(struct vpnproxy_data *proxy, zdtun_pkt_t *pkt, zdt
 static void check_tls_mitm(zdtun_t *tun, struct vpnproxy_data *proxy, zdtun_pkt_t *pkt, zdtun_conn_t *conn) {
     conn_data_t *data = zdtun_conn_get_userdata(conn);
 
-    if(shouldIgnoreApp(proxy, data->uid))
+    if(shouldIgnoreConn(proxy, zdtun_conn_get_5tuple(conn), data))
         return;
 
     if(pkt->tuple.ipproto == IPPROTO_TCP) {
@@ -657,7 +665,7 @@ static int connection_dumper(zdtun_t *tun, const zdtun_5tuple_t *conn_info, conn
     vpnproxy_data_t *proxy = (vpnproxy_data_t*) zdtun_userdata(tun);
     JNIEnv *env = proxy->env;
 
-    if(shouldIgnoreApp(proxy, data->uid)) {
+    if(shouldIgnoreConn(proxy, conn_info, data)) {
         /* Continue */
         return 0;
     }
@@ -843,8 +851,6 @@ static int run_tun(JNIEnv *env, jclass vpn, int tapfd, jint sdk) {
             .vpn_dns = getIPv4Pref(env, vpn, "getVpnDns"),
             .public_dns = getIPv4Pref(env, vpn, "getPublicDns"),
             .incr_id = 0,
-            .uid_filter = getIntPref(env, vpn, "getPcapUidFilter"),
-            .capture_unknown_app_traffic = (bool) getIntPref(env, vpn, "getCaptureUnknownTraffic"),
             .java_dump = {
                 .enabled = (bool) getIntPref(env, vpn, "dumpPcapToJava"),
             },
@@ -970,7 +976,11 @@ static int run_tun(JNIEnv *env, jclass vpn, int tapfd, jint sdk) {
                     goto housekeeping;
                 }
 
-                check_dns_req_dnat(&proxy, &pkt, conn);
+                if((check_dns_req_dnat(&proxy, &pkt, conn) == 0)
+                        && (pkt.tuple.dst_ip == proxy.vpn_dns)) {
+                    log_android(ANDROID_LOG_DEBUG, "ignoring packet directed to the virtual DNS server");
+                    goto housekeeping;
+                }
 
                 if(proxy.tls_decryption.enabled)
                     check_tls_mitm(tun, &proxy, &pkt, conn);
@@ -995,7 +1005,8 @@ housekeeping:
             sendCaptureStats(&proxy);
             proxy.capture_stats.new_stats = false;
             proxy.capture_stats.last_update_ms = now_ms;
-        } else if((now_ms - last_connections_dump) >= CONNECTION_DUMP_UPDATE_FREQUENCY_MS) {
+        } else if(((now_ms - last_connections_dump) >= CONNECTION_DUMP_UPDATE_FREQUENCY_MS) || dump_connections_now) {
+            dump_connections_now = false;
             sendConnectionsDump(tun, &proxy);
             last_connections_dump = now_ms;
         } else if((proxy.java_dump.buffer_idx > 0)
@@ -1058,5 +1069,10 @@ Java_com_emanuelef_remote_1capture_CaptureService_runPacketLoop(JNIEnv *env, jcl
                                                               jobject vpn, jint sdk) {
 
     run_tun(env, vpn, tapfd, sdk);
-    close(tapfd);
+}
+
+JNIEXPORT void JNICALL
+Java_com_emanuelef_remote_1capture_CaptureService_askConnectionsDump(JNIEnv *env, jclass clazz) {
+    if(running)
+        dump_connections_now = true;
 }