+++ app/src/main/jni/vpnproxy-jni/vpnproxy.h
@@ -74,8 +74,6 @@ typedef struct vpnproxy_data {
     vpn_conn_t *notif_pending;
     u_int32_t cur_notif_pending;
     u_int32_t notif_pending_size;
-    int uid_filter;
-    bool capture_unknown_app_traffic;
     uint64_t now_ms;
 
     struct {